Move the publish course filter form into the Courses namespace
Course reads as a member and Courses as a collection, and this form
filters a collection. Renames the sibling params class the same way so
the pair stays symmetrical, joining Publish::Courses::Query.

The class name is not part of any URL or param contract - the panel
renders with scope: "" - so there is no locale or request-format
fallout.
