+++ app/services/publish/courses/filter_params.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Publish
+  module Courses
+    # The query string params the publish course list accepts, mirroring
+    # Find::SearchParams. Each filter group is multi-select, so each arrives as an
+    # array.
+    class FilterParams
+      PERMITTED = [FilterForm::GROUPS.index_with { [] }].freeze
+
+      def self.permit(params)
+        params.permit(*PERMITTED)
+      end
+    end
+  end
+end
